VSCODE - Improve how the VSCode writer works and handles crashes
Previously there were typing issues causing crashes and format
exceptions. Now the code is a bit cleaner to follow and understand
especially in terms of handling named configs.

Wrapped writeConfig logic in a try/catch block to handle possible
comment failures and provide information for any crashes.

If there is a crash, it returns the original file content.